RavenDB-22429 Move CloudEventConverter to QueueEtl
