+++ src/Raven.Server/Documents/ETL/Providers/Queue/AzureQueueStorage/AzureQueueStorageEtl.cs
@@ -145,48 +145,4 @@ public sealed class AzureQueueStorageEtl : QueueEtl<AzureQueueStorageItem>
                 $"Failed to deliver message, Azure error code: '{ex.ErrorCode}', error reason: '{ex.Message}'", ex);
         }
     }
-
-    private sealed class CloudEventConverter : JsonConverter<CloudEvent>
-    {
-        public static readonly CloudEventConverter Instance = new CloudEventConverter();
-
-        const string SpecVersionAttributeName = "specversion";
-
-        private CloudEventConverter()
-        {
-        }
-
-        public override void Write(Utf8JsonWriter writer, CloudEvent cloudEvent, JsonSerializerOptions options)
-        {
-            writer.WriteStartObject();
-
-            writer.WritePropertyName(SpecVersionAttributeName);
-            writer.WriteStringValue(cloudEvent.SpecVersion.VersionId);
-
-            foreach (var pair in cloudEvent.GetPopulatedAttributes())
-            {
-                var attribute = pair.Key;
-                if (attribute == cloudEvent.SpecVersion.DataContentTypeAttribute ||
-                    attribute.Name == Partitioning.PartitionKeyAttribute.Name)
-                {
-                    continue;
-                }
-
-                var value = attribute.Format(pair.Value);
-
-                writer.WritePropertyName(attribute.Name);
-                writer.WriteStringValue(value);
-            }
-
-            writer.WritePropertyName("data");
-            writer.WriteRawValue(((BlittableJsonReaderObject)cloudEvent.Data).ToString());
-
-            writer.WriteEndObject();
-        }
-
-        public override CloudEvent Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            throw new NotImplementedException();
-        }
-    }
 }
